backend/requirements.txt
fastapi==0.95.1
openai==0.27.6
pandas==2.0.1
python-decouple==3.8
scipy==1.10.1
tiktoken==0.3.3
uvicorn==0.22.0
